Revert "Feat/migrate to bun (#135)"
This reverts commit 440cf302bcdaed45700053b32fe85d31fe044204.
