+++ apps/scraper/package.json
@@ -4,13 +4,13 @@
   "description": "A modern scraper for the uOttawa public course registry",
   "type": "module",
   "scripts": {
-    "build": "echo 'No Build needed for scraper with Bun'",
+    "build": "tsc",
     "check-types": "tsc --noEmit",
-    "db:cleanup": "bun src/cleanup.ts",
+    "db:cleanup": "node dist/cleanup.js",
     "lint": "eslint .",
-    "scrape:courses": "bun src/main.ts",
-    "scrape:subjects": "bun src/scrape/scrapeAvailableSubjects.ts",
-    "scrape:terms": "bun src/scrape/scrapeAvailableTerms.ts"
+    "scrape:courses": "node dist/main.js",
+    "scrape:subjects": "node dist/scrape/scrapeAvailableSubjects.js",
+    "scrape:terms": "node dist/scrape/scrapeAvailableTerms.js"
   },
   "dependencies": {
     "cheerio": "^1.1.2",
@@ -18,7 +18,8 @@
     "fetch-cookie": "^3.1.0",
     "neverthrow": "^8.2.0",
     "pino": "^10.1.0",
-    "pino-pretty": "^13.1.2"
+    "pino-pretty": "^13.1.2",
+    "puppeteer-core": "^24.29.1"
   },
   "devDependencies": {
     "@repo/db": "workspace:*",
@@ -27,7 +28,6 @@
     "@types/node": "^24.10.0",
     "domhandler": "^5.0.3",
     "eslint": "catalog:",
-    "puppeteer-core": "^24.29.1",
     "typescript": "catalog:"
   }
 }
