+++ apps/scraper/src/scrape.sh
@@ -1,21 +1,20 @@
 #!/bin/bash
 
-# Legacy method of scraping subjects
-# docker container run -d -p 9222:9222 zenika/alpine-chrome --no-sandbox --remote-debugging-address=0.0.0.0 --remote-debugging-port=9222 https://www.chromestatus.com/
-# bun scrape:subjects 
-# docker stop $(docker ps -q)
-# docker container prune -f
+# Still need to use Puppeteer for Subjects
+docker container run -d -p 9222:9222 zenika/alpine-chrome --no-sandbox --remote-debugging-address=0.0.0.0 --remote-debugging-port=9222 https://www.chromestatus.com/
+pnpm scrape:subjects 
+docker stop $(docker ps -q)
+docker container prune -f
 
 start_time=$(date +%s)
-bun scrape:subjects
-bun scrape:terms 
-bun scrape:courses 
+pnpm scrape:terms 
+pnpm scrape:courses 
 
 end_time=$(date +%s)
 execution_time=$((end_time - start_time))
 echo "Total Scraping Time: $execution_time seconds"
 
-bun db:cleanup
+pnpm db:cleanup
 
 # Take down AWS resources
 curl -L \
