+++ package.json
@@ -3,16 +3,6 @@
   "private": true,
   "license": "CC-BY-NC-SA-4.0",
   "author": "Raman Gupta",
-  "workspaces": {
-    "packages": [
-      "apps/*",
-      "packages/*"
-    ],
-    "catalog": {
-      "eslint": "^9.39.1",
-      "typescript": "^5.9.3"
-    }
-  },
   "scripts": {
     "build": "turbo run build",
     "check-types": "turbo run check-types",
@@ -23,13 +13,9 @@
     "format:packagejson": "sort-package-json ./**/package.json --ignore ./**/node_modules --ignore ./**/wasm",
     "lint": "turbo run lint"
   },
-  "dependencies": {
-    "minimatch": "^10.1.1"
-  },
   "devDependencies": {
     "@changesets/changelog-github": "^0.5.1",
     "@changesets/cli": "^2.29.7",
-    "@types/bun": "^1.3.2",
     "eslint": "catalog:",
     "prettier": "^3.6.2",
     "prettier-plugin-tailwindcss": "^0.7.1",
@@ -38,7 +24,7 @@
     "turbo": "^2.6.0",
     "typescript": "catalog:"
   },
-  "packageManager": "bun@1.3.2",
+  "packageManager": "pnpm@10.20.0+sha512.cf9998222162dd85864d0a8102e7892e7ba4ceadebbf5a31f9c2fce48dfce317a9c53b9f6464d1ef9042cba2e02ae02a9f7c143a2b438cd93c91840f0192b9dd",
   "engines": {
     "node": ">=22"
   }
