+++ vercel.json
@@ -5,8 +5,5 @@
       "changeset-release/main": false,
       "docs/*": false
     }
-  },
-  "installCommand": "curl -fsSL https://bun.sh/install | bash && ~/.bun/bin/bun install --frozen-lockfile",
-  "buildCommand": "cd ../../ && ~/.bun/bin/bun run build --filter @repo/env && ~/.bun/bin/bun run build --filter @repo/db && cd apps/web && ~/.bun/bin/bun --bun next build",
-  "bunVersion": "1.x"
+  }
 }
